- correct decl. of UA_Variant_copyRange() missing const - correct def of UA_Variant_decodeBinary() missing *UA_RESTRICT - no more Bug in void processJobs() jobsSize may be -1 - Timeout calculation in processRepeatedJobs() must be 32 Bit (if 16 bit overflow at 65 s) - some comments for UA_DateTime variables to check code - no "#ifdef _MSC_VER" in Subscription_registerUpdateJob thats stuff

+++ src/ua_types.c
@@ -171,7 +171,7 @@ int gettimeofday(struct timeval *tp, struct timezone *tzp) {
     SystemTimeToFileTime(&st, &ft);
     ul.LowPart  = ft.dwLowDateTime;
     ul.HighPart = ft.dwHighDateTime;
-    tp->tv_sec  = (ul.QuadPart - epoch) / 10000000L;
+    tp->tv_sec  = (long)((ul.QuadPart - epoch) / 10000000L);
     tp->tv_usec = st.wMilliseconds * 1000;
     return 0;
 }
